requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.6
django-filter==21.1
django-phone-verify==2.0.1
django-phonenumber-field==5.2.0
django-rest-framework==0.1.0
django-rest-sms-auth==0.1.22
djangorestframework==3.12.4
drfpasswordless==1.5.7
idna==3.2
kombu==5.1.0
nexmo==2.5.2
phonenumbers==8.12.29
Pillow==8.4.0
prompt-toolkit==3.0.19
pycodestyle==2.8.0
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
twilio==6.63.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
